Refactored dropoff page
